feat: add component props hook
